extract cardsize enum to own file
